perf(doctyped): separate builder to own file
